Move Automator and OSBrowser enums into SpireAutomator class. Fix bug where Safari browser was not recognized.
