--- src/spire/Automator.java
@@ -1,5 +1,0 @@
-package spire;
-
-public enum Automator {
-    ENROLLER, HOUSER
-}
